move nav files, remove mui
