+++ webapp/app.ironcalc.com/frontend/src/components/Navigation/navigation.css
@@ -1,0 +1,119 @@
+.nav-menu-wrapper {
+  position: relative;
+}
+
+.nav-menu-trigger {
+  display: flex;
+  align-items: center;
+  font-size: var(--typography-font-size);
+  font-family: var(--typography-font-family);
+  padding: 8px;
+  border-radius: 6px;
+  cursor: pointer;
+  background-color: transparent;
+  border: none;
+  color: var(--palette-common-black);
+}
+
+.nav-menu-trigger:hover,
+.nav-menu-trigger.is-active {
+  background-color: var(--palette-grey-100);
+}
+
+.nav-menu-panel {
+  position: absolute;
+  top: calc(100% + 4px);
+  left: -4px;
+  background: var(--palette-common-white);
+  border-radius: 8px;
+  padding: 4px 0;
+  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
+  z-index: 1300;
+}
+
+.nav-menu-item {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  font-family: var(--typography-font-family);
+  font-size: var(--typography-font-size);
+  width: calc(100% - 8px);
+  min-width: 172px;
+  margin: 0 4px;
+  border-radius: 4px;
+  padding: 8px;
+  height: 32px;
+  color: var(--palette-common-black);
+  gap: 8px;
+  border: none;
+  background: transparent;
+  cursor: pointer;
+  text-align: left;
+  box-sizing: border-box;
+  white-space: nowrap;
+}
+
+.nav-menu-item:hover {
+  background-color: var(--palette-grey-100);
+}
+
+.nav-menu-item svg {
+  width: 16px;
+  min-width: 16px;
+  height: 16px;
+  color: var(--palette-grey-600);
+}
+
+.nav-menu-item--delete {
+  color: var(--palette-error-main);
+}
+
+.nav-menu-item--delete svg {
+  color: var(--palette-error-main);
+}
+
+.nav-menu-item--delete:hover {
+  background-color: color-mix(
+    in srgb,
+    var(--palette-error-main) 10%,
+    transparent
+  );
+}
+
+.nav-menu-divider {
+  margin: 4px 0;
+  border: none;
+  border-top: 1px solid var(--palette-grey-200);
+}
+
+.nav-menu-item-text {
+  flex: 1;
+}
+
+.nav-menu-icon-placeholder {
+  width: 16px;
+  min-width: 16px;
+}
+
+/* Submenu — shown on hover of the parent item */
+.nav-menu-item--has-submenu {
+  position: relative;
+  justify-content: space-between;
+}
+
+.nav-menu-submenu {
+  display: none;
+  position: absolute;
+  left: 100%;
+  top: -4px;
+  background: var(--palette-common-white);
+  border-radius: 8px;
+  padding: 4px 0;
+  box-shadow: 0px 4px 12px rgba(0, 0, 0, 0.15);
+  z-index: 1301;
+}
+
+.nav-menu-item--has-submenu:hover .nav-menu-submenu,
+.nav-menu-item--has-submenu:focus-within .nav-menu-submenu {
+  display: block;
+}
